loader refactor, loaders package creation, loader type creation, loader factory creation
