+++ samoa-api/src/main/java/org/apache/samoa/evaluation/BasicRegressionPerformanceEvaluator.java
@@ -1,9 +1,5 @@
 package org.apache.samoa.evaluation;
 
-import java.util.List;
-
-import org.apache.samoa.instances.Attribute;
-
 /*
  * #%L
  * SAMOA
@@ -25,8 +21,7 @@ import org.apache.samoa.instances.Attribute;
  */
 
 import org.apache.samoa.instances.Instance;
-import org.apache.samoa.instances.Utils;
-import org.apache.samoa.moa.AbstractMOAObject;
+        import org.apache.samoa.moa.AbstractMOAObject;
 import org.apache.samoa.moa.core.Measurement;
 import org.apache.samoa.moa.core.Vote;
